backend/requirements.txt
flask==3.0.3
pywebview==5.2
pyinstaller==6.10.0
werkzeug==3.0.3
